pyproject.toml
[project]
name = 'rebalancinator'
version = '0.0.1'
dependencies = [
    'pyyaml >= 6.0, < 7.0',
]
